requirements.txt
flask==2.3.3
flask-smorest
flask-sqlalchemy
python-dotenv
marshmallow
gunicorn
psycopg2